CU-869ahw0mw: Add argument to control data flow when saving results.
When  is provided, the user (probably) expects the data to be saved on disk upon method call.
But the current implementation forced the user to iterate over the results to force the annotation to actually happen.
So this change allows the method to materialise the list internally to force the annotation to happen and results to be saved on disk.
Additionally, it adds 2 other options:
1. The lazy iteration (what happens when no  is provided) where the iteration of data is left to the user
2. The combined / saved and return option where the results are materialised, but also yielded. Notably, this will take up a lot of memory if/when used with large data sets