CU-869ahw0mw: Fix issue withe multiprocessing.
The previous implementation would always consider the method a generator. And as such, the work would never be done
at call time, regardless of whether or not the  was provided. This commit fixes that by making the
wrapper method a regular method that (sometimes) returns the iterator and other times just a (potentially empty) list.