CU-869awf45h: Update patch release script for more flexibility.
This will allow a minor release after a pre-release to be pushed out without any cherry-picked cahnges.
Normally, with a patch release, you would expect to cherry-pick at least something, but with a pre-release
that is being spun out as a full release, that's not always the case.

So this PR will allow no cherry-picks only if the version being created is a minor release after a pre-release for
said minor release.

PS: The naming of the 2 release scripts isn't great. The prepare_patch_release.sh script is designed to deal
with anything that already had a release within the same minor release cycle (i.e 2.2.1 after 2.2.0, but also 2.2.0 after 2.2.0a1 or 2.2.0rc1)