CU-869awy4ux: Set pretend version before installation of package.
This should fix the installation picking up (and trying to infer a version) from tags unrelated to medcat-den.

Currently, the workflow is picking up trainer's medcat-trainer/v3.1.0 tag and getting confused since that does not work for the regex (which expected medcat-den/v*).
And as such, the workflow is failing. See for example:
https://github.com/CogStack/cogstack-nlp/actions/runs/18660829454/job/53200931476