CU-869bj8g9k: Fix hardcoded requirement for spacy model download.
So far, if there was no Internet access, the fallback spacy model download failure would raise an exception and stall the entire process. Most models should come with their own spacy model anyway (if that's what they use). So the fallback model shouldn't be needed most of the time.

So this PR allows the subprocess for spacy model download to fail if there's a network issue. This should make it easier to use the library in scenarios where this method is called. This normally happens if/when a model is created from scratch and no on-disk model is provided. But it can alsoaffects converting models from v1 to v2 format.